Updated Final Writeup
I changed the image signatures to HTML tags, and I reworked some of the
grammer / vocab.  Plus I think the music is a good addition so I put it
as a link in the text near the top of the page.   Added at the top of
the page MAE6226 and the date it was finished.